Introduce PaginatorInterface
Introduced SimplePaginator, too.
Cache & Filter is required getItemsByPage
