+++ src/FilterTrait.php
@@ -1,0 +1,46 @@
+<?php
+/**
+ * Zend Framework (http://framework.zend.com/)
+ *
+ * @link      http://github.com/zendframework/zf2 for the canonical source repository
+ * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   http://framework.zend.com/license/new-bsd New BSD License
+ */
+
+namespace Zend\Paginator;
+
+use Zend\Filter\FilterInterface;
+
+trait FilterTrait
+{
+
+    /**
+     * Result filter
+     *
+     * @var FilterInterface
+     */
+    protected $filter = null;
+
+    /**
+     * Get the filter
+     *
+     * @return FilterInterface
+     */
+    public function getFilter()
+    {
+        return $this->filter;
+    }
+
+    /**
+     * Set a filter chain
+     *
+     * @param  FilterInterface $filter
+     * @return Paginator
+     */
+    public function setFilter(FilterInterface $filter)
+    {
+        $this->filter = $filter;
+
+        return $this;
+    }
+}
